LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/error-type-3-expected.txt

CONSOLE MESSAGE: [object Object]
CONSOLE MESSAGE: SyntaxError: Importing binding name 'default' cannot be resolved by star export entries.

PASS instantiation error has higher priority than evaluation error
